# Project_web1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Project_web1, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/To_Do_": {
      "get": {
        "tags": [
          "To_Do_"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/To_Do"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/To_Do"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/To_Do"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "To_Do_"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/To_Do"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/To_Do"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/To_Do"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/To_Do_/{id}": {
      "get": {
        "tags": [
          "To_Do_"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/To_Do"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/To_Do"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/To_Do"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "To_Do_"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "To_Do": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "subjectId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}